# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submit:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a new SPMB application
      description: Stores student registration data and the uploaded supporting documents.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/StudentApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"
        "400":
          description: Invalid form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request or uploaded file is too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many submissions from the same network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /submissions/status:
    get:
      operationId: getSubmissionStatus
      tags: [applications]
      summary: Check a public SPMB application status
      description: Looks up the current application status using the number printed on the registration receipt.
      parameters:
        - name: number
          in: query
          required: true
          schema:
            type: string
            pattern: "^(SPMB-)?[0-9]+$"
          description: Application number, for example SPMB-000001.
      responses:
        "200":
          description: Current application status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionStatusResponse"
        "400":
          description: Invalid application number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Application number not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    committeeSession:
      type: apiKey
      in: cookie
      name: spmb_committee_session
  schemas:
...
    StudentApplicationInput:
      type: object
      required:
        - jenjang
        - nama_calon
        - nama_panggilan
        - jenis_kelamin
        - tempat_lahir
        - tanggal_lahir
        - nik_anak
        - alamat_domisili
        - anak_ke
        - jumlah_saudara
        - status_anak
        - agama
        - warga_negara
        - tinggi_badan
        - berat_badan
        - transportasi
        - jarak_sekolah
        - nomor_kk
        - nik_ayah
        - nik_ibu
        - nomor_hp_orangtua
        - email
        - nama_ayah
        - ttl_ayah
        - pendidikan_ayah
        - pekerjaan_ayah
        - penghasilan_ayah
        - instansi_jabatan_ayah
        - nama_ibu
        - ttl_ibu
        - pendidikan_ibu
        - pekerjaan_ibu
        - penghasilan_ibu
        - instansi_jabatan_ibu
        - foto_3x4
        - akte_lahir
        - kartu_keluarga
        - ktp_orangtua
        - bukti_bayar
      properties:
        jenjang: { type: string, enum: [Playgroup, Daycare, TK-A, TK-B, SD, SMP] }
        nama_calon: { type: string, minLength: 2, maxLength: 100 }
        nama_panggilan: { type: string, minLength: 1, maxLength: 100 }
        jenis_kelamin: { type: string, enum: [Laki-laki, Perempuan] }
        tempat_lahir: { type: string, minLength: 1, maxLength: 100 }
        tanggal_lahir:
          type: string
          format: date
          description: "Usia dihitung pada 1 Juli 2027. Minimum: Playgroup 3 tahun, TK-A 4 tahun, TK-B 5 tahun, SD 6 tahun. Daycare dan SMP tidak memiliki batas usia minimum."
        nisn: { type: string, pattern: "^[0-9]{10}$", maxLength: 10, nullable: true }
        nik_anak: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        alamat_domisili: { type: string, minLength: 1, maxLength: 1000 }
        anak_ke: { type: string, pattern: "^[0-9]+$", description: "Integer from 1 to 20" }
        jumlah_saudara: { type: string, pattern: "^[0-9]+$", description: "Integer from 0 to 50" }
        status_anak: { type: string, minLength: 1, maxLength: 50 }
        agama: { type: string, minLength: 1, maxLength: 50 }
        warga_negara: { type: string, minLength: 1, maxLength: 50 }
        tinggi_badan: { type: string, pattern: "^[0-9]+([.][0-9]+)?$", description: "Height in centimeters" }
        berat_badan: { type: string, pattern: "^[0-9]+([.][0-9]+)?$", description: "Weight in kilograms" }
        riwayat_penyakit: { type: string, maxLength: 1000, nullable: true }
        transportasi: { type: string, minLength: 1, maxLength: 50 }
        jarak_sekolah: { type: string, minLength: 1, maxLength: 100 }
        nama_sekolah_asal: { type: string, minLength: 1, maxLength: 255, nullable: true, description: "Required for SD and SMP; not required for Playgroup, Daycare, TK-A, or TK-B" }
        tahun_lulus: { type: string, pattern: "^[0-9]{4}$", nullable: true, description: "Graduation year, 1900 or later; required for SD and SMP" }
        alamat_sekolah_asal: { type: string, minLength: 1, maxLength: 1000, nullable: true, description: "Required for SD and SMP; not required for Playgroup, Daycare, TK-A, or TK-B" }
        nomor_kk: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        nik_ayah: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16 }
        nik_ibu: { type: string, pattern: "^[0-9]{16}$", minLength: 16, maxLength: 16, nullable: true }
        nomor_hp_orangtua: { type: string, minLength: 10, maxLength: 16 }
        email: { type: string, pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$", maxLength: 254 }
        nama_ayah: { type: string, minLength: 1, maxLength: 100 }
        ttl_ayah: { type: string, minLength: 1, maxLength: 150 }
        pendidikan_ayah: { type: string, minLength: 1, maxLength: 100 }
        pekerjaan_ayah: { type: string, minLength: 1, maxLength: 100 }
        penghasilan_ayah: { type: string, minLength: 1, maxLength: 100 }
        instansi_jabatan_ayah: { type: string, minLength: 1, maxLength: 150 }
        nama_ibu: { type: string, minLength: 1, maxLength: 100 }
        ttl_ibu: { type: string, minLength: 1, maxLength: 150 }
        pendidikan_ibu: { type: string, minLength: 1, maxLength: 100 }
        pekerjaan_ibu: { type: string, minLength: 1, maxLength: 100 }
        penghasilan_ibu: { type: string, minLength: 1, maxLength: 100 }
        instansi_jabatan_ibu: { type: string, minLength: 1, maxLength: 150 }
        nama_wali: { type: string, maxLength: 100, nullable: true }
        hubungan_wali: { type: string, maxLength: 100, nullable: true }
        foto_3x4: { type: string, description: "Uploaded JPG or PNG file, maximum 5 MB" }
        akte_lahir: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        kartu_keluarga: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        ktp_orangtua: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
        bukti_bayar: { type: string, description: "Uploaded PDF, JPG, or PNG file, maximum 5 MB" }
...
    SubmissionResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        id: { type: number }
        nis: { type: string, nullable: true }
        receiptUrl: { type: string }
      required: [success, message, id, nis, receiptUrl]
    SubmissionStatusResponse:
      type: object
      properties:
        id: { type: number }
        applicationNumber: { type: string }
        nama_calon: { type: string }
        jenjang: { type: string }
        status: { type: string, enum: [Baru, Diverifikasi, "Perlu Perbaikan", Diterima, Ditolak] }
        created_at: { type: string }
      required: [id, applicationNumber, nama_calon, jenjang, status, created_at]
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        fields:
          type: array
          items: { type: string }
      required: [error]